requirements-dev.txt
# Development dependencies
-r requirements.txt

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Code quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
